--- funstruct/__init__.py
@@ -1,2 +1,0 @@
-from funstruct._frozendict import frozendict
-from funstruct._cons import CList, Cons, Nil
